[LibOS/libs/golang] Go binary patching
These patches add an extension library to LibOS which supports binary
patching of Go language applications, replacing low-level functions in
the Go runtime with wrappers that redirect to LibOS functionality.

It is enabled by adding the library to an application's manifest file as
a preloaded library.

This library depends on the symbol table extension library.

See the README for more information.

Signed-off-by: Alex Merritt <mail@alexmerritt.us>